Leap interface and Avatar: left-hand IK works for Leap interface.
This is in and smooth, and there's a decision to be made:
In the case where only one hand is visible, we have to decide if it's the left or right hand. Currently it's simple, based on which side it's detected on.
So far this seems to work best, in the (very) common case where hands go in/out of the sensor's field frequently.
We could also use a history-based heuristic to hold one hand as left/right, but that means that when it guesses wrong the error lingers.
Detecting handedness based on where the thumbs are would be better, but is unreliable on the Leap.
I went for the simple/clear version, but we can do whatever we want.