Update behavior of chat window close button
Use a stylesheet to suppress highlighting of the background when
pressing the button down.  Instead, it will change the image.

Unfortunately I wasn't able to find a way to do this without requiring
an extra image, or by doing it through the icon system.  The different
icon states didn't seem to take effect for button presses, and using
opacity in the stylesheet doesn't work yet for `QPushButton`s.