Add chat window dragging
This was originally implemented by Mohammed Nafees, shared on Basecamp.

I've noticed an issue with dragging across screens (monitors) where the
chat window will jump vertically.  I've tried debugging it, and
ultimately it looks like an issue with the `move` method.  The
coordinates are valid (I tried using global coordinates as well), and
the amount to move is correct, but when moving to coordinates on a
different screen it doesn't do the right thing.