Fix issue with pref dialog disappearing
With the frameless dialog being updated to support different types of
window, the pref dialog was no longer able to rely on it to keep itself
open in certain cases.  The window flags were also updated to update
after a "browse" call rather than in the slot so that it would still
execute in the case of cancelled selection.