Job #19700 BUG: Crash in NodeBounds::draw() fixed.
QMap and QReadWriteLock are orthogonal. Combined them into a wrapper
class called NodeToJurisdictionMap, replacing typedef. This allows us to
avoid changing method signatures wherever NodeToJurisdictionMap is used.
The lock is bound with the map and is available to all clients of the
NodeToJurisdictionMap. The lock allows multi-threaded access to the map.

Fixed compiler warning in NodeBounds.cpp regarding loss of precision
during conversion of double to GLfloat.