This patch updates Application::updateLocationInServer()  in Application.cpp
The online status is updated when the position is send to the data web, a avatar gets an offline status if the data web did not get an update in the last 30 seconds.
This function had an LastLocationObject in which only send in the location info to the data web in case the location changed.

Removed the last location so the position of the avatar is send to the data web every 5 seconds, so that the online status remains online if the avatar is still online and does not switch to offline when the avatar does not move.