Adding followThroughOverlappingAction script
The purpose of this project is to validate the principles of animation
into a procedural animation script. To test the theory, we implemented a
simple procedural arm swing animation that is both realtime and that
incorporates Disney's principles of animation (principle of Overlapping
Action and Follow Through). When applied to the arm swing, the
principles make the shoulder rotation lead the elbow rotation which will
lead the wrist rotation. Doing so results in an organic and life-like
effect.