converting existing JS scripts to use new damping
Old formula the scripts were using:  v *= damping

So a value of 0.999 was "little damping" and a value of 0.001 was "very strong damping",
but now that is reversed.

New formula: v *= (1 j- damping)^dt

A damping value of 0.001 now means "very little damping"
and a value of 0.999 means "very strong damping".