Interview project updates.
Tried to address all of Brad's notes, most of which were related to
matching the coding style for the project.  Also used GeometryCache
instead of making direct calls to OpenGL to do drawing, took a different
approach to seeding rand(), updated the packet version, and fixed a bug
that I noticed in the setting of the dimensions for the particle effect
entity.