The 3Dconnextion files from https://github.com/highfidelity/hifi/pull/5351
For now without a merge conflict.

Updated the menu name.

Still have to look at the fast zooming and yaw on windows, probably have to add a var to prevent the button changes to be pushed to fast. Not sure why the yaw thing does not always work, could be that the position is also send at the same time and the input mapper does not not process all those synchronical. Probably will have to do something with masking the postion when the rotation is set for yaw.