Fix for ATI discrete graphics on 2015 macbook pro / mac pro.
Fixed incorrect size for camera transform uniforms passed via
glBindBufferRange.   This is a legitimate bug on our side.
I guess Intel and nVidia drivers just happen to ignore this field.
This was likely causing many of the shader TransformCamera members
to be identity or un-initialized, specifically, the projection
matrix values.

This was resulting in a black screen on ATI gpus.