Fix for RenderableParticleffectEntityItem assert.
When a particle system had no living particles, PendingChanges::updateItem
would assert, due to the access to _vertices[0].  In release this might be
harmless as this memory is never accessed, but different implementations of
std::vector might do different things here.  So, let's be safe and early
return when the number of particles is 0.