Change to boom offset for 3rd person HMD camera.
Previously the boom offset was computed in HMD sensor space instead of
world space, so it did not play well with the vive, or the oculus
if you weren't facing the camera directly.

Now it is computed in world space, and is always behind the character's
position/orientation.  This can cause the boom to swing if the character
rotation changes but now the avatar rotation in HMD uses comfort mode.
It's not that disorienting.