Fixes for Mac build and crash on startup
For some reason the qt code gen is getting confused by the #if 0
in 3DConnextionClient.h, so I added a stub implementation.

In Application.cpp the change of moving idle into paintGL had a sideeffect
of calling OffscreenGlCanvas before it was initialized.  To work around this
I added a guard to prevent calling idle before all the gl windows/widgets have
been initialized.