Fix render rate.
Depending on how the threads get assigned, the actual render rate (not just the stats) could be as little as a third of what is targetted.

That's because the code was insisting on an empty texture escrow before our update/render cycle could begin, and otherwise just waits for the next whole cycle.
On my MacbookPro, this was regularly giving a render average of 20fps instead of 60.