Remove light curve scale/bias from cutoff
Light intensity attenuation is still calculated
as a function of distance from the light surface,
but is biased back so that the curve begins at
the light source, not the surface.

Light maximum radius is still user-inputted, but
if the radius is after a MIN_CUTOFF_INTENSITY,
a new cutoff radius will be calculated from the
intensity to avoid extra shader processing.