Grab.js: Update Position on Mouse Grab Start
Fixes grab event issue where if the object is moving, the position does
not update until the user has moved their mouse.

This commit makes sure that the moment the user starts dragging on a
object that isnt grabbed by another, the position is immediately updated
instead of waiting for the user to move their mouse.