Avatar: update renderItems when avatar position or render mesh is in the frustum.
Previously, when render mesh was visible the renderItems they were not updated with the latest avatar mixer joints rotations and position.
This would result in a frozen avatar being visible.

Now the renderItems are updated when the render mesh is in the frustum or the bounding sphere around the avatar position is visible.