away.js: simplify kneel animation handling
Instead of manually setting anim vars to play the kneel animation
and enable/disable IK, use the MyAvatar.overrideAnimation() and
MyAvatar.restoreAnimtion() functions, which did not exist when
away.js was written.  This should make it more robust and eliminate
the issue with IK being stuck in the disabled state.