Avatar eye gaze fix for HMDs
When no avatars are chosen for more detailed eye gaze tracking the eyes should look straight ahead.
Before this fix, if you were in room scale or not aligned with the camera sensor your gaze could drift to the left or to the right of your head's forward direction.
Now it should follow your head more closely.