Fix address bar becoming unusable after going backwards or forwards
The address bar dialog was being disabled when a lookup request was finished,
but was not hidden - this was bug #1. Presumably, though, the intent was to have the
window _not_ disappear when going backwards or forwards, but hiding the
window when a lookup request finishes would cause the window
to be hidden when going backwards or forwards - this was bug #2.

I decided to disable the hide-on-request-finished functionality because
the window is already manually hidden after entering a new address.