AvatarMixer: Clients will show incompatible version dialog
For this to work, the server needs to send an empty AvatarIdentity packet back to the sender
when it receives a packet mismatch error.  This AvatarIdentity packet will be different then
what the client expects and will trigger the incompatible version dialog.

Previously, the avatar-mixer was just silently dropping incoming mismatched version packets.
Causing the client to never get a response, and thus never showing the incompatible version dialog.