Support for scale property on model overlays
If "scale" property is present but "dimensions" are not, the model is NOT scale to fit.
And the scale of the model's natural dimensions will be affected by the scale properties.

If only the "dimensions" property is present, the model will "scale to fit" the dimensions.

If both properties are present, the model still be "scale to fit" but the dimension will be scaled by the scale factor.

For example:

If a model is loaded that is 2cm tall, is loaded with no "dimensions" or "scale" properties.
It will be displayed as 2cm tall.

    {"scale": 2}

The above properties will result in a model that is 4cm tall.

    {"dimensions": 1}

This will result in a model that is 1cm tall.

    {"scale": 2, "dimensions" 2}

Will result in a model that is 2cm tall.