far-grabbed objects no longer fly away when the avatar body shifts
Before this fix, a hand vector was being tracked that was the difference between the avatar's hand in avatar space with the avatar's hand
in avatar space the previous frame.  This hand vector was used to move the grabbed object position.  However, when the body shifts, objects
in the avatar's space will change rapidly, this would cause this hand vector to be incorrect and cause the object to shoot off in the distance.

Now, we track this hand delta in sensor a.k.a. room space.  This is immune to the shifts caused by body shifting.