Snapshot button plus script
Minor change to allow snapshots to not notify (though the old way
with CTRL+S still does).  Added button to do it, saves them to disk.

The plan is for us to add UI to share (or not) the snapshot.  That's
why we are not going through the notifications.  Also, the
script makes sure to hide/unhide hud and overlays.  Next we will upload
the pick to AWS via data-web (if you are logged in), and _then_ make
the share UI.