Bug fix for click events received by entity scripts
PointerEvent.isLeftButton should be true even on left button release events.
in a previous PR this behavior was changed. isLeftButton was used as a flag indicating the button state, which would be false on left button release events.
Because we have scripts that rely on the old behavior, I've changed it back to the original and introduced isPrimaryHeld properties instead.