Prevent crash on exit.
If you or another user begins to near-grab an object while you exit the application, you might run into this crash.
The PhysicalEntitySimulation has a list of AvatarHoldActions shared_ptr's.
When it is destroyed, via the Application destructor, it will release these pointers.
This in turn will invoke a AvatarHoldAction destructor, which attempts to reference the AvatarManager via the DependencyManager.
However at this point the AvatarManager has already been destroyed, leading to a null dereference.

The fix is to check the AvatarManager pointer first.