Heartbeats need to happen while waiting for sandbox
We can only know if sandbox is up by waiting for the http request to
succeed or fail.  In the common case (fail since it isn't up yet), we
have to wait until a connect timeout in http.  Seems that is a couple
seconds, and in that time we really upset the deadlock watchdog thread
since it keeps a moving average of heartbeat times.

Simple solution, updateHeartbeat while waiting, and right before we
start.