First cut
Started with the horrible screaming when avatar goes away, seems
due to not putting a codec string in the packets.  This was an existing
issue, not due to recent changes.  Also, some weird indenting was
fixed, etc...  Still hear an artifact when audio starts, but only when
a codec was negotiated.  Hoping to fix that too.