Fix for crash in openvr on startup
pluginActivate can fail and return false, when steamvr is in a "Not Ready" state.
In practice this can occur on a botched steamvr install, or when the vr compositor process is unable to start.

When pluginActivate returns false, there's some sort of race condition where pluginUpdate will still be called.
But it's only reproducible in release builds, without the debugger attached.

This commit prevents pluginUpdate from crashing by making sure that the openvr _system ptr is valid first.