Tablet can switch to a web url.
Help button has been updated to switch to the help.html.  It doesn't look good because
the content was authored expecting landscape window, not portrait.

The API is by no means final.  But it exposes two methods to JS.

* Tablet.gotoHomeScreen() - Used to forcefully go back to the home tablet "springboard", eventually this will
  hooked up to the tablet button.
* Tablet.gotoWebScreen(url) - Causes the tablet to switch from the home "springboard" and load a web url.

tabletTest.js and help.js have been updated to use this functionality.