Throttle extra avatar data
When PAL is open, we send AvatarData for avatars out of view, or
those who have ignored us (if we are an admin).  For those out of
view, we really don't need their info at 45hz, so this is a super
simple way to get it less often.  Chose 1/10th (4.5 hz) randomly
and could be much lower.  Just wanted to push this and get the
conversation going on how best to do it.