Bugfix, plus minor cr stuff
Seems to solve the 'due north' issue - we now have the main overlay as
the sphere, and it is the only thing clickable.  The model is no longer
clickable, so you cannot click clouds to select things.  Prevents some
confusion when opaque objects have ignoreRayIntersection, allowing you
to select a cloud you cannot see (i.e clicking near but not on the
sphere you can see).