HMD hover working now
Had to hack a bit -- the trigger mappings were not firing so instead I
just get the hand that is triggering (when HMD is active) and handle the
hover almost as before, with mouse move events.  Some thought about how
to map a 'mouse move' event into something where 2 triggers can both be
pressed at same time was in order.