Fix Overlay-as-child of model entity with reg point
This fixes as issue where when you set the parent of an overlay to be a
model entity with a non-(0.5, 0.5, 0.5) registration point, the overlay would
still be positioned relative to the center of the model. The issue is
that the default parent joint index was 0, which is the center of the
model.

Entities had a default parent joint index of -1 (technically 65535
because it is a uint16). Avatars previously defaulted to 0, and that is
maintained here.