Fix for findEntities not returning some entities that are within range.
This bug has been around a long time... introduced by commit 39ed7f7b in 2014.
This became apparent to me when testing the tablet-ui, there was a case where the
hand styluses were not appearing when close to a web entity.  Because findEntities
is such a fundamental feature used by handControllerGrab, this fix should make all
grabbing feel more consistent and predictable.

I added a unit test that reproduces the issue.