Fix problem with ContextMenu triggering on button up and button down.
This is fixed by clearing the ButtonPressedMap in the KeyboardMouseDevice every update.
Previously, this button would never be cleared, causing other controller actions mapped to the same button to have the incorrect state.
In this case, ContextMenu action would become 2 when pressed and 1 when released, anytime after the right mouse button was pressed.