21356 - Reorder Edit.js
Made changes to the 3 sub files for the properties tabs. Although
visually the exiting menu looked like it was in different sections,
structurally it was all just one run of item after item, with a fair
amount of div-itis. I added in fieldsets to provide true groups and
replaced much of the div-ities with aditional feidlsets as they are the
smeantic element to use here.