Get marketplaceID into imported entities
Rather than doing this on the backend, we can do it here at the
time we import.  Later we will have some support for getting
an entities identity hash and querying the backend to get a
marketplaceID for them as needed.  However for now this is enough
to get us moving forward.