WORKLIST ISSUE # 21471 : FIX PARENT JOINT IN EDIT JS
In entityPropeties.js
line #1079 says:
elParentJointIndex.addEventListener('change', createEmitNumberPropertyUpdateFunction('parentJointIndex'));

The problem should be fixed by changing it to:
elParentJointIndex.addEventListener('change', createEmitNumberPropertyUpdateFunction('parentJointIndex', 0));

createEmitNumberPropertyUpdateFunction() is parsing the value as a
float to a default decimal position of 4. Looks like it only accepts
integers. By adding in that 0 as a second argument, it should coerce the
value to one that will be recognized by the UI/backend.